Revert "Better text box"
