+++ easygui/boxes/demo.py
@@ -20,11 +20,11 @@ print(__package__ is None)
 
 import utils as ut
 
-from button_box import buttonbox
+from base_boxes import buttonbox
 from text_box import textbox
-from diropen_box import diropenbox
-from fileopen_box import fileopenbox
-from filesave_box import filesavebox
+from base_boxes import diropenbox
+from base_boxes import fileopenbox
+from base_boxes import filesavebox
 
 from derived_boxes import ynbox
 from derived_boxes import ccbox
