+++ easygui/boxes/derived_boxes.py
@@ -7,22 +7,12 @@
 """
 
 
-
-
-
 # Refs:
 #   https://www.python.org/dev/peps/pep-0366
 #   http://stackoverflow.com/questions/11536764/attempted-relative-import-in-non-package-even-with-init-py
-
-
 if __name__ == "__main__" and __package__ is None:
     from os import path, sys
     sys.path.append(path.dirname(path.abspath(__file__)))
-
-from multi_fillable_box import __multfillablebox
-from fillable_box import __fillablebox
-from button_box import buttonbox
-from choice_box import __choicebox
 import base_boxes as bb
 import text_box as tb
 import utils as ut
@@ -158,7 +148,7 @@ def boolbox(msg="Shall I continue?", title=" ",
             'boolbox takes exactly 2 choices!  Consider using indexbox instead'
         )
 
-    reply = buttonbox(msg=msg,
+    reply = bb.buttonbox(msg=msg,
                          title=title,
                          choices=choices,
                          image=image,
@@ -189,7 +179,7 @@ def indexbox(msg="Shall I continue?", title=" ",
       which button should be pressed
     :return: the index of the choice selected, starting from 0
     """
-    reply = buttonbox(msg=msg,
+    reply = bb.buttonbox(msg=msg,
                          title=title,
                          choices=choices,
                          image=image,
@@ -225,7 +215,7 @@ def msgbox(msg="(Your message goes here)", title=" ",
         raise AssertionError(
             "The 'ok_button' argument to msgbox must be a string.")
 
-    return buttonbox(msg=msg,
+    return bb.buttonbox(msg=msg,
                         title=title,
                         choices=[ok_button],
                         image=image,
@@ -369,7 +359,7 @@ def multenterbox(msg="Fill in values for the fields.", title=" ",
     :param list values: a list of field values
     :return: String
     """
-    return __multfillablebox(msg, title, fields, values, None)
+    return bb.__multfillablebox(msg, title, fields, values, None)
 
 
 # -----------------------------------------------------------------------
@@ -414,7 +404,7 @@ def multpasswordbox(msg="Fill in values for the fields.",
         print("Reply was: %s" % str(fieldValues))
 
     """
-    return __multfillablebox(msg, title, fields, values, "*")
+    return bb.__multfillablebox(msg, title, fields, values, "*")
 
 
 # -------------------------------------------------------------------
@@ -444,7 +434,7 @@ def enterbox(msg="Enter something.", title=" ", default="",
     :return: the text that the user entered, or None if he cancels
       the operation.
     """
-    result = __fillablebox(
+    result = bb.__fillablebox(
         msg, title, default=default, mask=None, image=image, root=root)
     if result and strip:
         result = result.strip()
@@ -463,7 +453,7 @@ def passwordbox(msg="Enter your password.", title=" ", default="",
     :return: the text that the user entered, or None if he cancels
       the operation.
     """
-    return __fillablebox(msg, title, default, mask="*",
+    return bb.__fillablebox(msg, title, default, mask="*",
                             image=image, root=root)
 
 
@@ -489,7 +479,7 @@ def multchoicebox(msg="Pick as many items as you like.", title=" ",
 
     global __choiceboxMultipleSelect
     __choiceboxMultipleSelect = 1
-    return __choicebox(msg, title, choices)
+    return bb.__choicebox(msg, title, choices)
 
 
 # -----------------------------------------------------------------------
@@ -510,7 +500,7 @@ def choicebox(msg="Pick something.", title=" ", choices=()):
 
     global __choiceboxMultipleSelect
     __choiceboxMultipleSelect = 0
-    return __choicebox(msg, title, choices)
+    return bb.__choicebox(msg, title, choices)
 
 
 # -----------------------------------------------------------------------
