+++ easygui/boxes/state.py
@@ -21,6 +21,3 @@ TEXT_ENTRY_FONT_SIZE = 12  # a little larger makes it easier to see
 
 
 STANDARD_SELECTION_EVENTS = ["Return", "Button-1", "space"]
-
-WIDTH_TEXT_PROP = 60
-WIDTH_TEXT_FIXED = 80
