+++ easygui/boxes/text_box.py
@@ -175,7 +175,7 @@ def textbox(msg="", title=" ", text="", codebox=0):
 
     # this line MUST go before the line that destroys boxRoot
     areaText = textArea.get(0.0, 'end-1c')
-    boxRoot.quit()
+    boxRoot.destroy()
     return areaText  # return __replyButtonText
