+++ easygui/boxes/updatable_text_box.py
@@ -10,13 +10,7 @@ Version |release|
 import sys
 import string
 
-# Refs:
-#   https://www.python.org/dev/peps/pep-0366
-#   http://stackoverflow.com/questions/11536764/attempted-relative-import-in-non-package-even-with-init-py
-if __name__ == "__main__" and __package__ is None:
-    from os import path
-    sys.path.append(path.dirname(path.abspath(__file__)))
-import utils as ut
+from . import utils as ut
 
 if sys.hexversion >= 0x020600F0:
     runningPython26 = True
@@ -257,7 +251,7 @@ def textbox(msg="", title=" ", text="", codebox=0, get_updated_text=None):
 
     # this line MUST go before the line that destroys boxRoot
     areaText = textArea.get(0.0, 'end-1c')
-    boxRoot.quit()
+    boxRoot.destroy()
     return areaText  # return __replyButtonText
