+++ easygui/easygui.py
@@ -93,10 +93,10 @@ if __name__ == "__main__" and __package__ is None:
     from os import path, sys
     sys.path.append(path.dirname(path.abspath(__file__)))
 # Import all functions that form the API
-from boxes.button_box import buttonbox
-from boxes.diropen_box import diropenbox
-from boxes.fileopen_box import fileopenbox
-from boxes.filesave_box import filesavebox
+from boxes.base_boxes import buttonbox
+from boxes.base_boxes import diropenbox
+from boxes.base_boxes import fileopenbox
+from boxes.base_boxes import filesavebox
 
 from boxes.text_box import textbox
 
@@ -118,7 +118,6 @@ from boxes.derived_boxes import multchoicebox
 from boxes.egstore import EgStore
 
 from boxes.about import eg_version, egversion, abouteasygui
-
 
 from boxes.demo import egdemo
