--- exportToHTML/multiprocess_prototype_C.py.html
@@ -1,113 +1,0 @@
-<html>
-<head>
-<title>multiprocess_prototype_C.py</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<style type="text/css">
-.ln { color: rgb(0,0,0); font-weight: normal; font-style: normal; }
-.s0 { }
-.s1 { color: rgb(0,128,0); font-weight: bold; }
-.s2 { color: rgb(0,0,128); font-weight: bold; }
-.s3 { color: rgb(128,128,128); font-style: italic; }
-.s4 { color: rgb(0,0,255); }
-</style>
-</head>
-<BODY BGCOLOR="#ffffff">
-<TABLE CELLSPACING=0 CELLPADDING=5 COLS=1 WIDTH="100%" BGCOLOR="#C0C0C0" >
-<TR><TD><CENTER>
-<FONT FACE="Arial, Helvetica" COLOR="#000000">
-multiprocess_prototype_C.py</FONT>
-</center></TD></TR></TABLE>
-<pre>
-
-<span class="s0">__author__ = </span><span class="s1">'Robert'</span><span class="s0"> 
-__version__ = </span><span class="s1">'prototype C'</span><span class="s0"> 
-</span><span class="s2">import </span><span class="s0">sys 
- 
-sys.path.append(</span><span class="s1">'..'</span><span class="s0">) 
-</span><span class="s2">import </span><span class="s0">easygui </span><span class="s2">as </span><span class="s0">eg 
- 
-</span><span class="s3">###############################</span><span class="s0"> 
-</span><span class="s2">print </span><span class="s1">&quot;The next three examples all do the same thing.  I gave all three just to show&quot; </span><span class="s0">\ 
-      </span><span class="s1">&quot;what the code would look like&quot;</span><span class="s0"> 
- 
-</span><span class="s3"># This works just like before</span><span class="s0"> 
-reply = eg.buttonbox()  </span><span class="s3"># This blocks until the user presses a button or closes the window.</span><span class="s0"> 
-</span><span class="s2">print </span><span class="s0">reply 
- 
-</span><span class="s3"># *** This is the new stuff ***</span><span class="s0"> 
-</span><span class="s3"># Horst, I am trying to make the old and new things use exactly the same functions (like buttonbox).  Is this a</span><span class="s0"> 
-</span><span class="s3">#        mistake?  Should I create a new API?</span><span class="s0"> 
- 
-my_box2 = eg.buttonbox(run=False)  </span><span class="s3"># Create a new button box.  Immediately returns.</span><span class="s0"> 
-</span><span class="s1">&quot;&quot;&quot; 
-    :param str msg: the msg to be displayed 
-    :param str title: the window title 
-    :param list choices: a list or tuple of the choices to be displayed 
-    :param str image: Filename of image to display 
-    :param str default_choice: The choice you want highlighted when the gui appears 
-    :param str cancel_choice: If the user presses the 'X' close, which button should be pressed 
-    :param bool run: If True, dialog opens, blocks, and then returns information pressed.  Default is True 
-                     If False, dialog opens and immediately returns with a pointer to that dialog 
-    :param bool visible: If True, dialog is immediately visible.  Only matters if run is False.  default == True 
-    :return: if run==True: Returns the text of the button that the user selected 
-             if run==False: Returns a button box dialog that can be used later 
-&quot;&quot;&quot;</span><span class="s0"> 
- 
-</span><span class="s3"># Change to have four buttons</span><span class="s0"> 
-my_box2.choices = [</span><span class="s1">'Yes'</span><span class="s0">, </span><span class="s1">'No'</span><span class="s0">, </span><span class="s1">'Maybe'</span><span class="s0">, </span><span class="s1">&quot;I don't care&quot;</span><span class="s0">] 
- 
-</span><span class="s3"># Disable fourth option</span><span class="s0"> 
-my_box2.choices[</span><span class="s4">3</span><span class="s0">].enabled = False 
- 
-</span><span class="s3"># Put these images on the four buttons</span><span class="s0"> 
-my_box2.choices.images = [</span><span class="s1">'check.jpg'</span><span class="s0">, </span><span class="s1">'x.jpg'</span><span class="s0">, </span><span class="s1">'question.jpg'</span><span class="s0">, </span><span class="s1">'XXXX.jpg'</span><span class="s0">] 
- 
-</span><span class="s3"># Change the text of the first two buttons</span><span class="s0"> 
-my_box2.choices[</span><span class="s4">0</span><span class="s0">].text = </span><span class="s1">&quot;Yup&quot;</span><span class="s0"> 
-my_box2.choices[</span><span class="s4">1</span><span class="s0">].text = </span><span class="s1">&quot;Nope&quot;</span><span class="s0"> 
-my_box2.msg = </span><span class="s1">&quot;Start pressing buttons!!!!&quot;</span><span class="s0"> 
-my_box2.images = [</span><span class="s1">'foo.jpg'</span><span class="s0">, </span><span class="s1">'bar.jpg'</span><span class="s0">, </span><span class="s1">'baz.jpg'</span><span class="s0">, </span><span class="s1">'bog.jpg'</span><span class="s0">] 
-my_box2.timeout = </span><span class="s4">10  </span><span class="s3"># Timeout after 10 seconds</span><span class="s0"> 
-my_box2.timeout_return_value = None  </span><span class="s3"># If timeout is reached, return this value</span><span class="s0"> 
- 
-</span><span class="s3"># Display my_box2 and count the number of times a button is pressed.  if the user does not press a button</span><span class="s0"> 
-</span><span class="s3"># with 10 seconds of previously pressing a button, time out and exit.</span><span class="s0"> 
-press_count = </span><span class="s4">0</span><span class="s0"> 
-</span><span class="s2">while </span><span class="s0">True: 
-    reply = my_box2.run() 
-    </span><span class="s2">if </span><span class="s0">reply </span><span class="s2">is </span><span class="s0">None: 
-        </span><span class="s2">break</span><span class="s0"> 
-    press_count += </span><span class="s4">1</span><span class="s0"> 
-    my_box2.msg = </span><span class="s1">&quot;You have pressed a button {0} time(s)&quot;</span><span class="s0">.format(press_count) 
-    </span><span class="s2">print </span><span class="s0">reply 
- 
-</span><span class="s3"># User has either hit cancel, or the timeout has triggered</span><span class="s0"> 
-my_box2.hide() 
- 
- 
-</span><span class="s3"># Let's show that dialog again and create another one.  Then &quot;run&quot; them both.  The return 'reply' is a special</span><span class="s0"> 
-</span><span class="s3"># instance which sort of acts like a string and sort of acts like a complex object, so if you say 'print reply',</span><span class="s0"> 
-</span><span class="s3"># you get a simple return string.  But you can also say reply.widget and potentially other things such as type</span><span class="s0"> 
-</span><span class="s3"># of event thrown.</span><span class="s0"> 
-my_box2.show() 
-my_box3 = eg.buttonbox(run=False) 
-my_box3.choices = [</span><span class="s1">'Yup'</span><span class="s0">, </span><span class="s1">'Nope'</span><span class="s0">] 
-press_count = </span><span class="s4">0</span><span class="s0"> 
-</span><span class="s2">while </span><span class="s0">True: 
-    reply = eg.run_all() 
-    </span><span class="s2">if </span><span class="s0">reply </span><span class="s2">is </span><span class="s0">None: 
-        </span><span class="s2">break</span><span class="s0"> 
-    press_count += </span><span class="s4">1</span><span class="s0"> 
-    </span><span class="s2">if </span><span class="s0">reply == </span><span class="s1">&quot;Yup&quot;</span><span class="s0">: 
-        </span><span class="s2">if </span><span class="s0">reply.widget == my_box2: 
-            </span><span class="s2">print</span><span class="s0">(</span><span class="s1">&quot;my_box2 said Yup&quot;</span><span class="s0">) 
-        </span><span class="s2">if </span><span class="s0">reply.widget == my_box3: 
-            </span><span class="s2">print</span><span class="s0">(</span><span class="s1">&quot;my_box3 said Yup&quot;</span><span class="s0">) 
-my_box2.hide() 
-my_box3.hide() 
- 
- 
- 
-</span></pre>
-</body>
-</html>
